deploy-package/backend/requirements.txt
# Windows：请用「标准」Python 3.13（python.exe，GIL 开启），不要用 python3.13t 建 venv，否则 pydantic-core 二进制与解释器 ABI 不匹配。
fastapi==0.115.2
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
PyYAML==6.0.2
httpx==0.27.2
PyJWT[crypto]==2.9.0
